timestamp/pyproject.toml
[tool.poetry]
name = "fcc-timestamp"
version = "0.1.0"
description = "freeCodeCamp.com Timestamp Microservice project"
authors = ["Daniel Lemm <61800298+ffe4@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
python-dateutil = "^2.8.1"

[tool.poetry.dev-dependencies]
pytest = "^5.4.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
